refactor: use u32 for block numbers

+++ starknet/src/types/proposal.cairo
@@ -3,13 +3,12 @@ use serde::Serde;
 use starknet::ContractAddress;
 use sx::types::{FinalizationStatus, UserAddress};
 
-/// NOTE: Using u64 for block numbers instead of u32 which we use in sx-evm. can change if needed.
 #[derive(Clone, Drop, Serde, PartialEq, starknet::Store)]
 struct Proposal {
-    snapshot_block_number: u64,
-    start_block_number: u64,
-    min_end_block_number: u64,
-    max_end_block_number: u64,
+    snapshot_block_number: u32,
+    start_block_number: u32,
+    min_end_block_number: u32,
+    max_end_block_number: u32,
     execution_payload_hash: felt252,
     execution_strategy: ContractAddress,
     author: UserAddress,
